Temporary workaround for AOT compiled VO literals.
Literals like (VO{:foo "bar"}) get interned at
macro time into a separate namespaces Var.
Unfortunately, Clojure is not smart enough to
add the value code to the static class
initializer, and there is no metadata or anything
to help the compiler…

For now nothing gets interned. :(

We probably should generate a class with a static
field or something instead of a Var.
Or figure out a way to add code to the class
initializer.